test: sched_perf: total number of queues option
Change -q option to be the total number of queues with events.
There's no need to limit the number of queues to a multiple of
worker thread count. Also, add an option for number of dummy
(empty) queues. This enables testing scheduler performance
when all created queues are not used.

Signed-off-by: Petri Savolainen <petri.savolainen@linaro.org>